Refactor mod keybind to only attach when modcard is open
